RAGGemini/requirements.txt
google-genai
python-dotenv
pymongo
pandas
openpyxl
sentence-transformers